LLVM : Update to version 15.0.7
This fixes some OSL build issues on Xcode 14.3.1 as well as brings us a little more up to date. LLVM 15.0.7 is the latest version supported by OSL 1.12 and is the same version used by Arnold 7.3.

To save a bit of time on CI, we're now only building the "Native" target, which targets the current system architecture only.